rename to vs-expandable
