+++ packages/vlossom/src/components/index.ts
@@ -6,7 +6,7 @@ export { default as VsBar } from './vs-bar/VsBar.vue';
 export { default as VsButton } from './vs-button/VsButton.vue';
 export { default as VsContainer } from './vs-container/VsContainer.vue';
 export { default as VsDivider } from './vs-divider/VsDivider.vue';
-export { default as VsExpandTransition } from './vs-expand-transition/VsExpandTransition.vue';
+export { default as VsExpandable } from './vs-expandable/VsExpandable.vue';
 export { default as VsFocusTrap } from './vs-focus-trap/VsFocusTrap.vue';
 export { default as VsFooter } from './vs-footer/VsFooter.vue';
 export { default as VsForm } from './vs-form/VsForm.vue';
