+++ packages/vlossom/src/components/vs-accordion/types.ts
@@ -1,6 +1,6 @@
 import type VsAccordion from './VsAccordion.vue';
 import type { BoxStyleSet, SizeStyleSet, TextStyleSet } from '@/declaration';
-import type { VsExpandTransitionStyleSet } from '@/components/vs-expand-transition/types';
+import type { VsExpandableStyleSet } from '@/components/vs-expandable/types';
 
 declare module 'vue' {
     interface GlobalComponents {
@@ -11,5 +11,5 @@ declare module 'vue' {
 export type { VsAccordion };
 
 export interface VsAccordionStyleSet extends SizeStyleSet, Omit<BoxStyleSet, 'display'>, TextStyleSet {
-    expand?: VsExpandTransitionStyleSet;
+    expand?: VsExpandableStyleSet;
 }
