--- packages/vlossom/src/components/vs-expand-transition/VsExpandTransition.css
@@ -1,59 +1,0 @@
-/* VsExpandTransition 기본 구조 스타일 */
-.vs-expand-transition-wrapper {
-    --vs-expand-transition-backgroundColor: initial;
-    --vs-expand-transition-padding: initial;
-
-    --vs-expand-transition-fontColor: initial;
-    --vs-expand-transition-fontSize: initial;
-    --vs-expand-transition-fontWeight: initial;
-    --vs-expand-transition-lineHeight: initial;
-    --vs-expand-transition-whiteSpace: initial;
-
-    @apply relative w-full overflow-hidden;
-
-    background-color: var(--vs-expand-transition-backgroundColor, var(--vs-area-bg));
-}
-
-.vs-expand-transition-content {
-    @apply w-full;
-
-    padding: var(--vs-expand-transition-padding, 0.75rem 1rem);
-    color: var(--vs-expand-transition-fontColor, var(--vs-comp-font));
-    font-weight: var(--vs-expand-transition-fontWeight, var(--vs-font-weight));
-    font-size: var(--vs-expand-transition-fontSize, var(--vs-font-size));
-    line-height: var(--vs-expand-transition-lineHeight, var(--vs-line-height));
-    white-space: var(--vs-expand-transition-whiteSpace, var(--vs-white-space));
-}
-
-/* VsExpandTransition 기본 애니메이션 스타일 */
-.vs-expand-enter-active,
-.vs-expand-leave-active {
-    transition:
-        height 0.3s cubic-bezier(0.4, 0, 0.2, 1),
-        opacity 0.6s ease-out;
-    overflow: hidden;
-}
-
-/* 확장 시작 상태 */
-.vs-expand-enter-from {
-    opacity: 0;
-    height: 0;
-}
-
-/* 확장 완료 상태 */
-.vs-expand-enter-to {
-    opacity: 1;
-    height: auto;
-}
-
-/* 축소 시작 상태 */
-.vs-expand-leave-from {
-    opacity: 1;
-    height: auto;
-}
-
-/* 축소 완료 상태 */
-.vs-expand-leave-to {
-    opacity: 0;
-    height: 0;
-}
