+++ packages/vlossom/src/components/vs-expandable/VsExpandable.css
@@ -1,0 +1,59 @@
+/* VsExpandable 기본 구조 스타일 */
+.vs-expandable-wrapper {
+    --vs-expandable-backgroundColor: initial;
+    --vs-expandable-padding: initial;
+
+    --vs-expandable-fontColor: initial;
+    --vs-expandable-fontSize: initial;
+    --vs-expandable-fontWeight: initial;
+    --vs-expandable-lineHeight: initial;
+    --vs-expandable-whiteSpace: initial;
+
+    @apply relative w-full overflow-hidden;
+
+    background-color: var(--vs-expandable-backgroundColor, var(--vs-area-bg));
+}
+
+.vs-expandable-content {
+    @apply w-full;
+
+    padding: var(--vs-expandable-padding, 0.75rem 1rem);
+    color: var(--vs-expandable-fontColor, var(--vs-comp-font));
+    font-weight: var(--vs-expandable-fontWeight, var(--vs-font-weight));
+    font-size: var(--vs-expandable-fontSize, var(--vs-font-size));
+    line-height: var(--vs-expandable-lineHeight, var(--vs-line-height));
+    white-space: var(--vs-expandable-whiteSpace, var(--vs-white-space));
+}
+
+/* VsExpandable 기본 애니메이션 스타일 */
+.vs-expand-enter-active,
+.vs-expand-leave-active {
+    transition:
+        height 0.3s cubic-bezier(0.4, 0, 0.2, 1),
+        opacity 0.6s ease-out;
+    overflow: hidden;
+}
+
+/* 확장 시작 상태 */
+.vs-expand-enter-from {
+    opacity: 0;
+    height: 0;
+}
+
+/* 확장 완료 상태 */
+.vs-expand-enter-to {
+    opacity: 1;
+    height: auto;
+}
+
+/* 축소 시작 상태 */
+.vs-expand-leave-from {
+    opacity: 1;
+    height: auto;
+}
+
+/* 축소 완료 상태 */
+.vs-expand-leave-to {
+    opacity: 0;
+    height: 0;
+}
